# Conflicts:
#	pom.xml
#	src/main/java/spring_server/chat_model/GameChatController.java
#	src/main/java/spring_server/chat_model/GameChatRoom.java
#	src/main/java/spring_server/chat_model/Message.java
#	src/main/java/spring_server/chat_model/Room.java
#	src/main/java/spring_server/chat_model/RoomsController.java
#	src/main/java/spring_server/chat_model/User.java
#	src/main/java/spring_server/controllers/GameChatRoomController.java
#	src/main/java/spring_server/controllers/RoomRequestController.java
#	src/main/java/spring_server/controllers/WebSocketChatController.java
#	src/main/java/spring_server/controllers/WebSocketChatEventListener.java

+++ src/main/java/spring_server/chat_model/RoomsController.java
@@ -10,9 +10,9 @@ public class RoomsController {
 
     private final int MAX_USERS = 5;
     @Getter
-    private final ConcurrentHashMap<Long, Room> gameRooms = new ConcurrentHashMap<>();
+    private final ConcurrentHashMap<Integer, Room> gameRooms = new ConcurrentHashMap<>();
 
-    private long createRoom() {
+    private int createRoom() {
         Room chatRoom = new Room("Room #" + gameRooms.size());
         gameRooms.put(chatRoom.getId(), chatRoom);
         return chatRoom.getId();
@@ -22,11 +22,11 @@ public class RoomsController {
         gameRooms.get(user.getRoomId()).removeUser(user.getId());
     }
 
-    public Room getRoomById(long id) {
+    public Room getRoomById(int id) {
         return gameRooms.get(id);
     }
 
-    public long addToRandomRoom(User user) {
+    public int addToRandomRoom(User user) {
         for (Room room : gameRooms.values()) {
             if (room.getUsers().size() < MAX_USERS) {
                 room.addUser(user);
@@ -35,7 +35,7 @@ public class RoomsController {
             }
         }
 
-        long roomId = createRoom();
+        int roomId = createRoom();
         gameRooms.get(roomId).addUser(user);
         user.setRoomId(roomId);
         user.setPainter(true);
