+++ src/main/java/spring_server/chat_model/User.java
@@ -14,7 +14,7 @@ public class User {
 
     @Getter
     @Setter
-    private long roomId;
+    private int roomId;
 
     @Getter
     private transient int score = 0;
